requirements.txt
# Framework web y utilidades
fastapi>=0.100.0          # Framework web moderno y rápido
uvicorn>=0.23.0          # Servidor ASGI para FastAPI
pydantic>=2.0.0          # Validación de datos y configuración
pydantic-settings>=2.0.3 # Configuración basada en variables de entorno

# Gestión de configuración
python-dotenv>=1.0.0     # Carga de variables de entorno desde .env

# Integración con IA
langchain>=0.0.330       # Framework para aplicaciones de IA
openai>=1.2.0            # Cliente oficial de OpenAI
langchain_community>=0.3.16  # Componentes comunitarios de Langchain

# Base de datos y autenticación
supabase>=2.0.0          # Cliente de Supabase
postgrest>=0.10.6        # Cliente PostgREST para Supabase
psycopg2-binary>=2.9.9   # Driver de PostgreSQL
gotrue~=2.11.3          # Cliente de autenticación de Supabase

# Utilidades
PyJWT>=2.8.0            # Manejo de JWT
requests>=2.31.0        # Cliente HTTP
python-multipart>=0.0.6 # Manejo de formularios multipart
typer>=0.9.0           # CLI moderno para Python
httpx>=0.24.0          # Cliente HTTP async